improved structure with ts
