+++ package.json
@@ -41,6 +41,7 @@
 	],
 	"dependencies": {
 		"react": "^18.2.0",
-		"react-dom": "^18.2.0"
+		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.9.0"
 	}
 }
