--- src/declarations/nexai/nexai.old.did
@@ -1,3 +1,0 @@
-service : {
-  greet: (text) -> (text);
-}
