+++ src/nexai_assets/src/app.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import { Routes, Route } from "react-router-dom";
+import Greetings from "../components/Greetings";
+
+const App = () => {
+	return (
+		<Routes>
+			<Route path='/' element={<Greetings />} />
+		</Routes>
+	);
+};
+
+export default App;
