+++ tsconfig.json
@@ -6,7 +6,8 @@
 			"DOM"
 		] /* Specify library files to be included in the compilation. */,
 		"allowJs": true /* Allow javascript files to be compiled. */,
-		"jsx": "react" /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
+		"jsx": "react" /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */,
+		"moduleResolution": "node"
 	},
 	"include": ["src/**/*"]
 }
